Strip Foreman check + recommend bundled Foreman
I think our current implementation here is a little split brain: we
bundle Foreman, but then check for a system version, and error when you
don't have it.

The current implementation of `bin/setup` doesn't necessitate Foreman
anywhere, so firstly, pull that check out. Then recommend the use of the
bundled Foreman in the README.